requirements.txt
torch
numpy
opencv-python
tqdm
matplotlib
Pillow
supervision
groundingdino-py
##  groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
git clone https://github.com/IDEA-Research/GroundingDINO.git
